pyproject.toml
[project]
name = "ps-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask-bcrypt>=1.0.1",
    "flask-sqlalchemy>=3.1.1",
    "jwt>=1.4.0",
]
